engine: Small speedups.
* Inline logN and popcnt
* Changed fix to take a pointer.

T:7500 @ TC:40/15+0.05
W:2182 L:2068 D:3250
LOS:0.960 ELO:5.3±5.9

+++ engine/misc.go
@@ -18,10 +18,10 @@ func logN(n uint64) uint {
 }
 
 const (
-	k1 = uint64(0x5555555555555555)
-	k2 = uint64(0x3333333333333333)
-	k4 = uint64(0x0f0f0f0f0f0f0f0f)
-	kf = uint64(0x0101010101010101)
+	k1 = 0x5555555555555555
+	k2 = 0x3333333333333333
+	k4 = 0x0f0f0f0f0f0f0f0f
+	kf = 0x0101010101010101
 )
 
 // popcnt counts number of bits set in n.
